Refactor task queue and request handling
- Updated TaskQueue class to allow dynamic database assignment and improved task handling.
- Changed request handling to use a class-based structure for better organization and maintainability.
- Updated various components to utilize the new database service import instead of global window.database.
- Refactored CloudDataApiDialog to use the new createService function for initializing requests.
