+++ src/api/cloud/gamelogJoinLeave.js
@@ -1,16 +1,16 @@
-export default {
-    insert: function (data) {
-        window.taskQueue.addTask({
-            url: '/game-log/join-leave',
-            method: 'POST',
-            data
-        });
-    },
-    batchInsert: function (data) {
-        window.taskQueue.addTask({
-            url: '/game-log/join-leave/batch',
-            method: 'POST',
-            data
-        });
-    }
-};
+import taskQueue from '../../service/taskQueue';
+
+export function insert(data) {
+    taskQueue({
+        url: '/game-log/join-leave',
+        method: 'POST',
+        data
+    });
+}
+export function batchInsert(data) {
+    taskQueue({
+        url: '/game-log/join-leave/batch',
+        method: 'POST',
+        data
+    });
+}
