+++ src/classes/gameLog.js
@@ -4,6 +4,7 @@ import gameLogService from '../service/gamelog.js';
 import configRepository from '../service/config.js';
 import { baseClass, $app, API, $utils } from './baseClass.js';
 import { userRequest } from '../api';
+import database  from '../service/database.js';
 import dayjs from 'dayjs';
 export default class extends baseClass {
     constructor(_app, _API, _t) {
@@ -157,7 +158,7 @@ export default class extends baseClass {
                         userId,
                         time: 0
                     };
-                    window.database.addGamelogJoinLeaveToDatabase(entry);
+                    database().addGamelogJoinLeaveToDatabase(entry);
                     break;
                 case 'player-left':
                     var ref = this.lastLocation.playerList.get(userId);
@@ -193,7 +194,7 @@ export default class extends baseClass {
                         userId,
                         time
                     };
-                    window.database.addGamelogJoinLeaveToDatabase(entry);
+                    database().addGamelogJoinLeaveToDatabase(entry);
                     break;
                 case 'portal-spawn':
                     if (this.ipcEnabled && this.isGameRunning) {
@@ -208,7 +209,7 @@ export default class extends baseClass {
                         instanceId: '',
                         worldName: ''
                     };
-                    window.database.addGamelogPortalSpawnToDatabase(entry);
+                    database().addGamelogPortalSpawnToDatabase(entry);
                     break;
                 case 'video-play':
                     gameLog.videoUrl = decodeURI(gameLog.videoUrl);
@@ -242,7 +243,7 @@ export default class extends baseClass {
                         resourceUrl: gameLog.resourceUrl,
                         location
                     };
-                    window.database.addGamelogResourceLoadToDatabase(entry);
+                    database().addGamelogResourceLoadToDatabase(entry);
                     break;
                 case 'screenshot':
                     // var entry = {
@@ -253,7 +254,7 @@ export default class extends baseClass {
                     //         ''
                     //     )}`
                     // };
-                    // window.database.addGamelogEventToDatabase(entry);
+                    // database().addGamelogEventToDatabase(entry);
 
                     this.processScreenshot(gameLog.screenshotPath);
                     break;
@@ -371,7 +372,7 @@ export default class extends baseClass {
                         type: 'Event',
                         data: gameLog.event
                     };
-                    window.database.addGamelogEventToDatabase(entry);
+                    database().addGamelogEventToDatabase(entry);
                     break;
                 case 'vrc-quit':
                     if (!this.isGameRunning) {
@@ -419,7 +420,7 @@ export default class extends baseClass {
                     //     type: 'Event',
                     //     data: gameLog.data
                     // };
-                    // window.database.addGamelogEventToDatabase(entry);
+                    // database().addGamelogEventToDatabase(entry);
                     break;
                 case 'sticker-spawn':
                     if (!$app.saveInstanceStickers) {
@@ -495,7 +496,7 @@ export default class extends baseClass {
                 ...input,
                 groupName
             };
-            window.database.addGamelogLocationToDatabase(entry);
+            database().addGamelogLocationToDatabase(entry);
         },
 
         async addGameLogVideo(gameLog, location, userId) {
@@ -890,7 +891,7 @@ export default class extends baseClass {
             if (this.gameLogTable.vip) {
                 vipList = Array.from(this.localFavoriteFriends.values());
             }
-            this.gameLogTable.data = await window.database.lookupGameLogDatabase(
+            this.gameLogTable.data = await database().lookupGameLogDatabase(
                 this.gameLogTable.search,
                 this.gameLogTable.filter,
                 vipList
@@ -928,7 +929,7 @@ export default class extends baseClass {
 
         // async refreshEntireGameLog() {
         //     await gameLogService.setDateTill('1970-01-01');
-        //     await window.database.initTables();
+        //     await database().initTables();
         //     await this.resetGameLog();
         //     var location = '';
         //     for (var gameLog of await gameLogService.getAll()) {
@@ -941,9 +942,9 @@ export default class extends baseClass {
         // },
 
         async getGameLogTable() {
-            await window.database.initTables();
-            this.gameLogSessionTable = await window.database.getGamelogDatabase();
-            var dateTill = await window.database.getLastDateGameLogDatabase();
+            await database().initTables();
+            this.gameLogSessionTable = await database().getGamelogDatabase();
+            var dateTill = await database().getLastDateGameLogDatabase();
             this.updateGameLog(dateTill);
         },
