+++ src/service/taskQueue.js
@@ -1,12 +1,14 @@
+import request from '../utils/requests';
+
 class TaskQueue {
-    constructor(db) {
-        this.db = db;
+    constructor() {
         this.queue = [];
         this.isProcessing = false;
     }
 
     // 初始化：加载数据库中未完成的任务
-    async init() {
+    async init(db) {
+        this.db = db;
         const pending = await this.db.getAllPendingTasks();
         this.queue.push(...pending);
         this._notify();
@@ -24,7 +26,9 @@ class TaskQueue {
             status: 'pending',
             created_at: Date.now()
         };
-        await this.db.insertTask(task);
+        if (this.db) {
+            await this.db.addTask(task);
+        }
         this.queue.push(task);
         this._notify();
     }
@@ -47,14 +51,18 @@ class TaskQueue {
         const task = this.queue.shift();
         try {
             await this._upload(task);
-            await this.db.markTaskSuccess(task.id);
+            if (this.db) {
+                await this.db.markTaskCompleted(task);
+            }
         } catch (e) {
             console.error('上传任务失败:', e);
             task.retry++;
             if (task.retry < 3) {
                 this.queue.push(task); // 重试
             } else {
-                await this.db.markTaskFailed(task);
+                if (this.db) {
+                    await this.db.markTaskFailed(task);
+                }
             }
         }
         this._processNext();
@@ -62,9 +70,16 @@ class TaskQueue {
 
     // 实际上传请求
     async _upload(task) {
-        await window.DAPI({
+        await request({
             ...task
         });
     }
 }
-export default TaskQueue;
+const toaskQueue = new TaskQueue();
+export function initTaskQueue(db) {
+    return toaskQueue.init(db);
+}
+
+export default function (data) {
+    return toaskQueue.addTask(data);
+}
